+++ app/src/components/MeetingView/MeetingView.tsx
@@ -1,8 +1,7 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { styled, Theme, useMediaQuery, Box } from '@mui/material';
-import { useTheme } from '@mui/material/styles';
+import { styled } from '@mui/material';
 import { useRef, useState, useEffect } from 'react';
 
 import { useAppSelector } from '../../hooks';
@@ -11,11 +10,9 @@ import { selectDebugMode, selectShowCoffeeBreakCurtain } from '../../store/slice
 import { selectIsModerator } from '../../store/slices/userSlice';
 import { CoffeeBreakView } from '../CoffeeBreakView/CoffeeBreakView';
 import DebugPanel from '../DebugPanel';
-import LocalVideo from '../LocalVideo';
 import RemoteAudioStreams from '../RemoteAudioStreams';
 import TimerPopover from '../TimerPopover';
-import Toolbar from '../Toolbar';
-import InnerLayout from './InnerLayout';
+import InnerLayout from './fragments/InnerLayout';
 
 const Container = styled('div')(({ theme }) => ({
   background: theme.palette.background.overlay,
@@ -28,33 +25,14 @@ const Container = styled('div')(({ theme }) => ({
     paddingLeft: 0,
     paddingRight: 0,
   },
-}));
-
-const ToolbarWrapper = styled(Box)(({ theme }) => ({
-  position: 'fixed',
-  bottom: 0,
-  zIndex: theme.zIndex.mobileStepper,
-  backgroundColor: theme.palette.background.paper,
-  width: '100%',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const LocalVideoWrapper = styled(Box)(({ theme }) => ({
-  position: 'fixed',
-  bottom: 65,
   [theme.breakpoints.down('sm')]: {
-    bottom: 87,
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'space-between',
   },
-  width: 100,
-  zIndex: theme.zIndex.mobileStepper,
 }));
 
 const MeetingView = () => {
-  const theme: Theme = useTheme();
-  const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'));
-  const isSmallPortraitScreen = useMediaQuery(`${theme.breakpoints.down('md')} and (orientation: landscape)`);
   const debugMode = useAppSelector(selectDebugMode);
   const isCoffeeBreakOpen = useAppSelector(selectShowCoffeeBreakCurtain);
   const isModerator = useAppSelector(selectIsModerator);
@@ -83,17 +61,6 @@ const MeetingView = () => {
           {!showCoffeeBreakCurtain && <TimerPopover anchorEl={anchorEl} />}
 
           <InnerLayout />
-
-          {(isSmallScreen || isSmallPortraitScreen) && (
-            <>
-              <LocalVideoWrapper>
-                <LocalVideo />
-              </LocalVideoWrapper>
-              <ToolbarWrapper>
-                <Toolbar />
-              </ToolbarWrapper>
-            </>
-          )}
         </>
       )}
     </Container>