+++ CHANGELOG.md
@@ -24,6 +24,10 @@
 - Fix handling with disallow custom display name ([!1089](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/merge_requests/1089))
 - Fix overflowing details on the meeting details page ([#1144](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1144))
 
+### Stability Improvements
+
+- Make meeting room header mobile responsive ([#1447](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1447))
+
 ## 1.6.0
 
 ### UI/UX Fixes